requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.4
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.15
django-cogwheels==0.2
django-cors-headers==3.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-el-pagination==3.2.4
django-htmlmin==0.10.0
django-modelcluster==4.3
django-mptt==0.8.7
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.6.4
draftjs-exporter==2.1.5
Faker==0.8.18
html5lib==0.999999999
idna==2.7
ipaddress==1.0.22
Markdown==2.6.11
Misago==0.19.5
misago-social-auth-app-django==2.1.0
oauthlib==2.1.0
olefile==0.45.1
path.py==10.3.1
Pillow==4.1.1
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
requests==2.20.0
requests-oauthlib==1.0.0
six==1.11.0
social-auth-core==1.7.0
sqlparse==0.2.4
text-unidecode==1.2
Unidecode==0.4.21
urllib3==1.23
wagtail==2.3
wagtailemoji==1.0.2
wagtailmenus==2.12
webencodings==0.5.1
Willow==1.1
